feat: :sparkles: New API
The new API introduces two new types `AdConnection` and `AdConnectionAsync`
which both have a `try_search()` method which will handling the searching of
LAPS passwords.

Those structs can be constructed via the `connect()` of `connect_async()`
function of the `AdSetting` struct.

This should simplify the API massively.

I'm not quite sure if this will be the final API or not